# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgencyOS API
servers:
  - url: /api
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: Clients list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /clients/{id}/contracts:
    get:
      operationId: getClientContracts
      tags: [clients]
      summary: Get client contracts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      responses:
        "200":
          description: Leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /leads/{id}:
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
  /invoices/{id}:
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List proposals
      responses:
        "200":
          description: Proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Create proposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    patch:
      operationId: updateProposal
      tags: [proposals]
      summary: Update proposal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
    delete:
      operationId: deleteProposal
      tags: [proposals]
      summary: Delete proposal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        isActive:
          type: ["boolean", "null"]
...
    Client:
      type: object
      required: [id, companyName]
      properties:
        id:
          type: string
        companyName:
          type: string
        contactPerson:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        health:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        serviceType:
          type: ["string", "null"]
        serviceDetails:
          type: ["string", "null"]
        socialHandles:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        contentFrequency:
          type: ["string", "null"]
        targetAudience:
          type: ["string", "null"]
    ClientInput:
      type: object
      required: [companyName]
      properties:
        companyName:
          type: string
        contactPerson:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        health:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        serviceType:
          type: ["string", "null"]
        serviceDetails:
          type: ["string", "null"]
        socialHandles:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        contentFrequency:
          type: ["string", "null"]
        targetAudience:
          type: ["string", "null"]
    Contract:
      type: object
      required: [id]
      properties:
        id:
          type: string
        title:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
    Lead:
      type: object
      required: [id, title, stage]
      properties:
        id:
          type: string
        title:
          type: string
        stage:
          type: string
        companyName:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        daysInStage:
          type: ["integer", "null"]
    LeadInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        stage:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
...
    Project:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        status:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
...
    Invoice:
      type: object
      required: [id]
      properties:
        id:
          type: string
        number:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        subtotal:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        total:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        lineItems:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/LineItem"
    InvoiceInput:
      type: object
      properties:
        clientId:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        subtotal:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        total:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        lineItems:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/LineItem"
    LineItem:
      type: object
      properties:
        description:
          type: string
        qty:
          type: number
        unitPrice:
          type: number
        taxPercent:
          type: number
...
    Proposal:
      type: object
      required: [id]
      properties:
        id:
          type: string
        title:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        template:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        validUntil:
          type: ["string", "null"]
        scope:
          type: ["string", "null"]
        deliverables:
          type: ["string", "null"]
        timeline:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ProposalInput:
      type: object
      properties:
        title:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        template:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        validUntil:
          type: ["string", "null"]
        scope:
          type: ["string", "null"]
        deliverables:
          type: ["string", "null"]
        timeline:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...