# lib/api-zod/src/generated/types/proposalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AgencyOS API
 * OpenAPI spec version: 0.1.0
 */

export interface ProposalInput {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  clientId?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  template?: string | null;
  /** @nullable */
  value?: number | null;
  /** @nullable */
  validUntil?: string | null;
  /** @nullable */
  scope?: string | null;
  /** @nullable */
  deliverables?: string | null;
  /** @nullable */
  timeline?: string | null;
  /** @nullable */
  notes?: string | null;
}